Add files via upload
I have no clue on how to share the dummy data from my PostgreSQL database, but I managed to create a dump file if that is of any use. 

Inside of it, you're supposed to see a database named geekText along with a table titled "library." Columns include book_id, author_name, book_name, genre, and rating. There are ten rows of data at the time of this creation, and will be subject to change possibly.